feat: add poor man's dot graph parser
This "parser" assumes the structure that is used to export dependency
graphs from bitbake. It uses very simple regular expressions and does
not deal with different type of node IDs, optional commas and
semicolons, chained edge statements, escaped double quotes inside of
quotes, comments, and probably more. The interface is generic enough to
replace the implementation later with a real parser if necessary.